Make sure the plugin names for PerfMon counters are stable.
There was a bit problem with the existing logic. If new Counter instances were added and removed (e.g. stopping/starting a website while monitoring IIS counters or dropping/adding a database in SQL server) then the counter "name" (the munin version) would potentially be reused for a different perf object.

I've changed this so that we always include the counter name itself in the munin name. This should be stable over the life of the object.